pyproject.toml
[tool.poetry]
name = "texturize"
version = "0.4.0"
description = ""
authors = ["Alex J. Champandard <445208+alexjc@users.noreply.github.com>"]
license = "AGPL-3.0"
packages = [
    { include = "texturize", from = "src" },
]

[tool.poetry.scripts]
texturize = 'texturize.__main__:main'

[tool.poetry.dependencies]
python = "^3.6"
creativeai = "^0.1"
docopt = "^0.6.2"
progressbar2 = "^3.51.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
